Make the installation of .stylelint more informative to the user

+++ includes/install/stylelintrc.json.sh
@@ -9,7 +9,16 @@ fi
 inform "Downloading $STYLELINT_SETTINGS..."
 curl -O https://gist.githubusercontent.com/code-warrior/a766f7c32bab9a82b467601800b00a46/raw/768717143df9db9c593dabb38c3c7aa63c87f66b/"$STYLELINT_SETTINGS"
 
-inform "Installing $STYLELINT_SETTINGS..."
-mv -v "$STYLELINT_SETTINGS" "$HOME/"
+if [[ -e "$STYLELINT_SETTINGS" ]]; then
+    inform "$STYLELINT_SETTINGS downloaded successfully."
 
-inform "$STYLELINT_SETTINGS downloaded and installed to $HOME/$STYLELINT_SETTINGS"
+    inform "Installing $STYLELINT_SETTINGS..."
+    mv -v "$STYLELINT_SETTINGS" "$HOME/"
+
+    if [[ -e "$STYLELINT_SETTINGS" ]]; then
+        warn "$STYLELINT_SETTINGS was not successfully installed. Please investigate, then continue."
+        pause_and_warn
+    else
+        inform "$STYLELINT_SETTINGS was installed successfully."
+    fi
+fi
